Handle stale VPN sessions
Now changes the display when a stale VPN session is detected:

It will look something like:
    <vpn?> HK

If you are connected to a VPN and let OSX sleep, sometimes the VPN
tunnel interface will still be "UP" (according to ifconfig) but actually
it is stale and a new connection is needed. We can detect this by
looking at the routing table, if the default gateway no longer points to
the tunnel interface.